implement voucher code retrieval
