+++ packages/frontend/.env.example
@@ -1,5 +1,6 @@
 NEXT_PUBLIC_INFURA_KEY=
-NEXT_VOUCHER_REDEEM_DEADLINE="2024-04-25T17:44:44.773Z"
+NEXT_PUBLIC_VOUCHER_REDEEM_DEADLINE="2024-04-25T17:44:44.773Z"
+VOUCHER_CODES=ba6c54fde086a987,4a90b64c32e2eec1,4456f133d558f072,c5baeed2585b3a1a,15e2e282fe03a9f6,d668bbbec3c7d3b6,f47d335b531ec0dd,1349f6e4ad95825b,c7de218efebac868,518ff50098b36b4f,5395a102b8deb36b,cd08815b9ae77ec8,df88717615180380,7eb8a8355ab38fd2,dd34debe1ece8929,9f27bc7014e0b8f6,2b5ab52d00c07988,f49da00473a58a0d,fcad48a66cc2f857,4d3ac5e8cf1fc722,4546f8288b02459a,a9a046b0755f0b35,2e28275b9e17a364,0c2e2baff9d6b044,6dbffb8bd731cc10,d0c4f3bd9a51c035,ea50061750e7220c,f62f668c8df8f033,ccdc220e6429e40b,b933753ac3065218,ad2741485309bb5e,a7cbdb7a436a393e,171fe26cff20e0c9,1a373416e1f6c65f,6d4907febab8d51e,ceb9aa23347b41e0,06c3439f8dfedf06,bcad460b53ad5af6,43b07ef8a50cdc0b,0bdd5b2d051f8c94,62ef486dcff2bb7d,bd43edb50dc78b42,d32c28cc390ae8c1,d76f404a01f3d365,297e3650df45ccee,854cdffbc1879804,5cb257d69ddbbeaa,2a8d7b0b19198740,8c1c78ffd189aae7,f51dc2b4460a80a3,fc5232592e0b90a7,dbc282570f659346,56dd89fa35a537a1,2c4771ae5018a794,0d9b5d1aab875f09,17489a24ec7404ae,e01c690f01f35f59,3161b7c9d6ebc3f2,90187ca640822f9c,c391985bfb80f179,793232c16f918943,0b4dd569e47746dc,2eed8b7899af6619,ce6fd19583df13ea,50fe857f53db54cc,7094b0d3d17da71e,e20525b54eab8049,358a07cf9507335f,686800d7853d4b64,40c0a7bcd91713d4,d9f26e0dea6a4ac2,d3475c5918b54546,75ccb2d3faad3202,66779020a6568504,a1805e90273f0887,0bd85a1c8ebbed9a,963cb6da9cf8d694,3b6b08cf4c30c277,654935b396b34ecc,2e3d1c016ac0e5b6,267f603bf10734aa,19b7d087960ef325,e5836a5d6f38bce5,3002fe977f618f12,ecc204e71ae0608a,81439b3700dac2e1,ed79494f356c9f56,f74d8afaf589b246,64c1b13df5f0e2a8,93fc83ed3343aefa,81494872a726ac5e,d51aafd59caf1d9a,b000e99d7f08d768,3689c0e9e4bfd7cc,d63c7a0354bbaf91,bbe027c85f54b22e,1b1a0b893326ab04,653e9f4bf4c0e49a,22126a9bae93c1d6,8567a17423a22df3,af639384ade8bba3,c7416b2579922cc3,da02ac1ac8532cac,e95feedd55f16e17,266d598e6c4ca76f,e11e0654cb8834f2,73f415136b33cbdf,d20744c053d33da0,ad6dbfe64f785f6d,e49c4b6cde6513b9,27f333f2ab79e8d8,827590ae7d70fdae,747bc37bafc0df75,186335bf4568b8dc,9840c5ae1ce08975,8ce15bd4fe2de3a0,896e12125cdd14f5,1772bf95537d9f92,45188d7df7948be4,3684339f00a9ee7d,38979cf3cba38b23,ecd87f3795c2164e,74636e2a1ed532e0,1b306af0faec308d,096074e8f4d52c34,fcc6a12f8289d2a9,039a8e8d80fa3834,064a0d9bf97a8806,4a5b93c5bcffda57,5fb463b3eb69ff99,d9583498d54fb835,be66e6f75f5190fc,5aaf4dab0369d01a,a18445e29f8e95bd,a36c622fe9857e66,902e74e0b287dacf,4f900eba5986aa87,fb455447611cd905,723e292ad57effab,01e2ac152929d3e7,de4fd43f8995fb11,e7731d83d20eb7e1,c865eea3cedb911d,2c45eef4be1affbf,782a91a43d1972bc,48d6a66cc950304e,e6f1242fc8d4ec84,508471b3ec37cebe,ab08f3ded5a76ba2,a5ec32f0e2949058,f9164e47f80cb3bd,bf38d15a84ee9ed1,6bd3b8424e506f00,0637ce97d7b9fe8d,b906af92ccd8b777,3b0fc5f260f42536,bafe611ce5f52a0f,19ec1a0671e98cf4,08e0326fe3b2a967,1069ae55f845a7c5,0e3f21740766115a,617aff575dc63e13,58911e28ca0885c8,74f7d9e5b21fa639,e7178bec1bac46f2,61ee32138d313e8f,ad2de280496b9bc2,dceb430e485375df,4e3e90630caba69c,d1bb15495477d718,fc0f9a3ea6c2a539,00dbea350f25bec1,fc598819c0879160,82a0c16147704ad0,610852b356b74e1f,acde744959fbdbf1,8e2c0e9f262d09c7,b20a509236cbb1dc,1b92b67e779cea96,a31949e13591119f,cdbb6313fda4fd7e,f78c434ceb60f7a7,01c2d74945179ae1,848841d938f158dc,bf544dd7c68b8086,f1a4d3ec90616adb,b86791be425ca33f,ff364c7529f4ec2b,ee59e0ea5628e1f6,3302ca90924171c0,b38fe194537ac74c,7d8035d532280e91,73c0e05518648e93,a8bd32595aea3a74,e49454aa99097d3b,ecac9c5d183fb029,e49f09faef030787,793d084a9763df9a,0af09e41f69e5238,c479d48056bd6857
 GTC_SCORER_API_BASE_URI=https://api.scorer.gitcoin.co
 GTC_SCORER_API_KEY="gitcoin scorer api key"
 GTC_SCORER_ID="gitcoin scorer id"
